+++ src/components/controllers/DatePickerController.tsx
@@ -2,7 +2,7 @@ import { Controller } from 'react-hook-form';
 import DatePickerInput from '@opetushallitus/virkailija-ui-components/DatePickerInput';
 import * as React from 'react';
 import { DayPickerProps } from 'react-day-picker/types/Props';
-import { ControllerProps } from '../types';
+import { ControllerProps } from '../../types';
 
 type Props<T> = ControllerProps<T> & {
     dayPickerProps?: DayPickerProps;
