+++ src/components/controllers/FastSelect.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { useMemo, useState } from 'react';
 import Select from '@opetushallitus/virkailija-ui-components/Select';
-import { SelectOption } from '../types';
+import { SelectOption } from '../../types';
 import { escapeRegExp, sortBy } from 'lodash';
 
 const MAX_DISPLAYED_OPTIONS = 500;
